improvement(rpm): add build.sh script to build contrib/fedora RPM
Run the script to build SRPM/RPM from the git checkout.

This script is also intended later to run in copr, to generate the
SRPM, for building Fedora/EPEL packages.

The firewalld.spec file cannot be used directly, instead the build
script will perform necessary text replacements before build.
This will mean, that the spec file cannot be directly used in Fedora,
but the places that require patching are minimal and obvious, so the
vast majority of the file will be usable.