improvement(rpm): introduce naming scheme for patches and rename patch
Introduce a naming scheme to differentiate between downstream patches
(which also apply after a rebase) and bugfixes. Further, also
differentiate between Fedora-only, RHEL-only and patches that apply to
Fedora/RHEL alike.

Yes, it's complicated. The complication comes from the (maybe ill
advised) desire to have different downstream behaviors for Fedora and
RHEL. Maybe that should be revisited, and the Fedora-only and RHEL-only
categories dropped. Other than that, this is just making sense of the
status quo.

This is an effort, to unify the spec files.