improvement(rpm): drop "Obsoletes: firewalld-selinux < 0.4.4.2-2" from contrib/fedora
This was added to Fedora 26 ([1]).

Note that the "Obsoletes" also found its way into rhel-8.0 (forked from
Fedora 28) and rhel-9.0. But neither rhel-8 nor rhel-9 ever had a
firewalld package as ancient as 0.4.4.

Note that rhel-7 never had a package "firewalld-selinux" either. So this
doesn't even affect users who upgrade from rhel-7.

Actually, "firewalld-selinux" subpackage only existed for a few days in
2016 (added in 0.4.4.1, dropped in 0.4.4.2-2).

Drop this. This change is also suitable for any recent RHEL builds.

[1] https://src.fedoraproject.org/rpms/firewalld/c/27c4930b69c5886aa304d82dc6ec082b009acf70

This is an effort, to unify the spec files.