fix(config): adjust service files to be valid according our XML schemas
For the XSD schemas, the order of XML elements is important. Maybe that
is a bug of the schemas, to be stricter than what firewalld actually
requires.

On the other hand, we only use the schemas for additional validation.
It's find to be extra strict there, while firewalld is more relaxed and
accepts XML elements in any order.

Adjust the service files.