improvement(nftables): rule_key: do not strip "index" et al.
This avoids a deepcopy().

Stripping "index" et al. is no longer necessary since we're not
recomputing the rule_key during rule application. As a consequence,
"index" is now part of the rule_key. This is okay for the hard coded
index locations. The generated index values are added _after_ the
rule_key is generated in set_rules().