style: reformat with black 23.7.0-2.fc39
Fedora 39 is out and ships black 23.7.0-2.fc39. There were some changes
in the formatting ([1]), which means that changes are introduced when
using black 23.

The officially blessed black version is still black=22.12.0, which you
can get via pip. This is also the version that the github action uses
for checking the formatting.

The good news is, once files are reformatted with black 23, this is fine
for both black 23 and black=22.12.0.

Do the reformatting, so that you can use black 23 just as well.

[1] https://black.readthedocs.io/en/stable/change_log.html#id38