improvement(rich): don't reset rule attributes in _import_from_string()
Note that _import_from_string() is only called from __init__().

The caller already ensures that all attributes are initialized, so there
is no need in doing that again.

Except, note that the caller sets `self.family` and `self.priority`, as
specified by the arguments of __init__(). It seems wrong to silently
overwrite those attributes.

In practice, this doesn't matter, because none of the callers set
"rule_str" together with "family"/"priority". So in all cases there is
no difference. This just drops unnecessary and surprising code.

In theory, it could make sense to specify "family"/"priority" together
with a "rule_str" argument, and that works now as expected (even if
unused). Specifying "priority" as argument to Rich_Rule(), will
overwrite the priority from "rule_str". Specifying "family" together
with "rule_str" will force the family to be as requested (or fail if the
family of the string disagrees).