improvement(dbus): use empty default signature argument in @dbus_service_method
@dbus.service.method has arguments in_signature and out_signature. They
default to None, which leaves the signature unspecified. We don't want
that. We want that we always specify the correct signature.

However, for convenience, allow the caller to omit the
in_signature/out_signature arguments, if the signature is empty. Thus
default them to "".

Note that dbus.service.method() will check that the in-signature matches
with the function arguments. This will blow up, if we wrongly didn't
specify a signature and the signature is not in fact empty.