improvement(dbus): extend @dbus_service_method decorator with an is_deprecated flag
Optimally, dbus-python would support annotations and/or an
"is_deprecated" argument. We should add that feature to the library, but
in any case, for now we need to work around it.

Simplify the workaround. Whether a method is deprecated is part of the
meta-data about that method. It should be all specified by one
decorator.

Also, since @dbus_service_method and @dbus_service_method_deprecated
both need an D-Bus interface, using two decorators means that the
information is redundant.