improvement(dbus): add is_deprecated argument to @dbus_service_signal decorator
Similar to @dbus_service_method, add a "is_deprecated" argument. This
will replace @dbus_service_signal_deprecated decorator.

Also, default "signature" to "". Almost all users of the
@dbus_service_signal decorator specify a signature argument. Those that
don't are for signals without arguments.  We should not leave the
signature unspecified. But for convenience, allow the user to omit the
signature if it's empty. The default argument helps with that.