improvement(config): add meta data for config properties
Add the list of known properties to a CONFIG_PROPERTIES dictionary. We
really should treat things more generically, where there is a list of
properties (and their parameters) that can be handled in a unified way.

This is all bad. It already starts that there is no concept (a function)
of how a boolean string looks like, instead Set() reimplements it as `if
new_value.lower() not in ["yes", "no", "true", "false"]`.

But in general, there is no list of the properties, their expected
types, how to parse, set, validate, serialize them. It's all
re-implemented one by one, and then not even using common functions like
a validate_string_as_bool() function.