chore(dbus_utils): remove unused functions
Maybe these functions should be kept, as being useful and valuable to have
in the toolbox, for the future.

Instead, drop them. It's unused, and it's not clear that they are the
best implementation we can have. When we need them again, we should consider
what the best implementation would be.