improvement(fw-in-container): don't do anything in "40-firewalld.rules" polkit rule
Messing with the polkit rules means that `make check-integration` will
fail. It should pass by default in fw-in-container.

Now the "40-firewalld.rules" file does nothing. It's still useful to
have, because whenever I want to edit the polkit rules for manual
testing, it is cumbersome to find the right location for the file and
the proper syntax. The "40-firewalld.rules" file helps with that.