improvement(nftables): simplify _rich_rule_priority_fragment() to add entry directly
It seems wasteful, to have _rich_rule_priority_fragment() return a dictionary,
and then call rule.update() on that result. Note that all callers do it exactly
the same way. Also note that most of the time, the priority is not present, and
the code wouldn't need to do anything else.

Rework the approach. The function is now called _rule_set_rich_rule_priority(),
and it sets the "%%RICH_RULE_PRIORITY%%" entry depending on the @rich_rule
argument.