fix(policy): dispatch update for active policies
For policies that are already active AND the ingress and egress zone
list contain zoneA, e.g. where ingress-zone=ANY and egress-zone=ANY,
then zoneA will NOT include the currently activating interface/source in
list_interfaces()/list_sources().

As such, we must manually add interface/source for the special case
where the zone is present in both the ingress and egress.

Fixes: #1278