improvement(functions): add str_to_bool() helper
All throughout the code base we re-implement how a boolean value may be
expressed. Add firewall.functions.str_to_bool() for *the* way to
convert a string to boolean. It accepts various alternative keywords,
strips whitespace and ignores letter case.

- By default, str_to_bool(val) is strict, and only accepts truthy/falsy
values (or raises an exception).

- With str_to_bool(val, fallback), the function never fails and returns
"fallback" for all unrecognized values.

- The function can also parse and distinguish a third value. If set to
None, "", "-1" or "default", then we can accept that as ternary. You get
that with

  - str_to_bool(val, on_default=def) (accepts default value, but is
    otherwise still strict.

  - str_to_bool(val, on_default=def, on_error=err) never fails, but
    allows you to distinguish between default/error.