improvement(firewalld_conf): normalize settings in firewalld_conf
In firewalld_conf, we validate the input. For example,

  if not value or value.lower() not in ["no", "false", "yes", "true"]:

However, no normalization happens. That means, the value can be set to
"No", "False" or "yes". That means, everybody down the road has to
perform the same validation of those strings. Which is worse, since this
is open coded here, so they have to duplicate the implementation of how
a boolean looks like.

Also, for example "IPv6_rpfilter" will default to "yes"/"no" due to

  self.set("IPv6_rpfilter", "yes" if config.FALLBACK_IPV6_RPFILTER else "no")

while "FlushAllOnReload" will default to "True"/"False" due to

  self.set("FlushAllOnReload", str(config.FALLBACK_FLUSH_ALL_ON_RELOAD))

Instead, normalize the values while validating them. Booleans get a
"yes"/"no" value. But we anyway should check booleans with
firewall.functions.str_to_bool() and not rely on this.

While at it, add helper functions to not repeat this over and over.

Also, for enums strip whitespace and allow values in upper case (but
they will be normalize to lower-case).

Also, booleans now use firewall.functions.str_to_bool(), which is the
new way for how a boolean string looks like. This also accepts values
like "1", "on" or "y", because, if a user configures such a value, their
intent is pretty clear and no point in being picky.