improvement(firewalld_conf): add generic approach for default/normalize
Instead of repeating the validation of each key over and over, add some
meta data (KeyType) for how to handle them generically.

Certain keys, like "ReloadPolicy",key_type=_parse_reload_policy are not
really re-used. However, those are the minority. We should instead aim
for simple properties (int, bool), which have a common behavior.

The default() and normalize() functions are not fully generic. Instead,
based on the key_type we switch on a hand full of behaviors.
Theoretically, we could go overboard and create various subclasses with
virtual functions. In practice, this approach is fine and pretty
extensible.

Note that firewalld_conf.set() always strips the string value. Normalize
does that too, and ensures that the value is always stripped.