improvement(functions): move _parse_reload_policy() function out of firewalld_conf class
The next commit will add key metadata, which is static (depending not on
an instance of firewalld_conf). While could be nested inside
firewalld_conf class, it's better to have it at module level.

That meta data will need the _parse_reload_policy() function. It's
awkward, if those are inside the firewalld_conf class. Move it.