improvement(nftables): make methods static that don't use self
Having a @self pointer begs the question, what state from the object we
use, and/or whether we modify the state of the object.

It is preferable to have less state, and have few access to state.
Making a method @staticmethod makes that clear and helps with review.

Also, drop the token argument and just hard code it to
"%%RICH_RULE_PRIORITY%%". Note that all other uses that refer to such
"%%" tokens also hard code them.