improvement(nftables): move json_cmd() to a separate helper function
We will rework the invocation of the JSON command, by writing the
JSON to file. For that, move the call to self._nft_ctx.json_cmd()
to a new helper.

Note that libnftables' json_cmd() is itself just a small wrapper
around json.dumps() and Nftables.cmd(). As we want to get close
to a place where we don't generate a JSON string in memory, take the
json_cmd() implementation and re-implement it.

Later we will modify it further.