improvement(nftables): write JSON to file for libnftables
Instead of passing the JSON to libnftables, write the JSON to a file in
/run/firewalld, and use cmd_from_file() to load it from there.

The ultimate goal is to reduce the amount of memory we keep around.

Note that the test don't create a separate mount namespace and don't
remount /run. They maybe should do that. But with this change, we
have yet another place where we write temporary files to /run. And if we
run tests in parallel, they could theoretically interfere. In practice
they don't, because the temporary files have unique names. In any case,
if this is a problem, the solution is that tests mount a private
/run/firewalld.