improvement(nftables): don't chunk set_rules() in set_restore()
The problem with chunking is that it splits the transaction, and what
should be an atomic update, is not. libnftables is supposed to provide
us with a concept of transactions, if we cannot use that because the
transaction is too large, then that's a fail.

We changed the format of the JSON for ipset entries, instead of an "add"
command for each entry, they are now all combed in one. This
significantly makes the operation faster, and we shouldn't need the
chunking anymore.

This in parts reverts 9a86f119e52a ('fix(ipset): chunk entries when
restoring set'), however, in the meantime it is thought that the
chunking is no longer necessary.