improvement(nftables): free json data earlier during nft_cmd()
The JSON data may be large, and we have to convert it to string.
Optimally, we free it as soon as we no longer need it.

Previously, if nft_cmd() failed, we would raise a ValueError() with the
full JSON blob as error message. That means, we had to keep the JSON
blob around for longer. It's questionable whether we really need this.
Note that if you run with  debug logging, then we log the full JSON before
making the call. If we don't have debug logging enabled, then we
shouldn't embed the full JSON data in our error message.

By not embedding the JSON string in the ValueError() after failure, we
can get rid of it earlier.