improvement(fw-in-container): support running nftables test in container
fw-in-container is useful for testing and development.

Note that it runs a rootless container. For non-root users, the netlink
socket buffer is limited by /proc/sys/net/core/{rmem_max,wmem_max}. This
causes various tests inside fw-in-container to fail. The test also
cannot just bump that limit, because it's rootless (and because it would
be an undesirable modification of the host).

However, it's useful to support running such tests also inside the
container. All that we need, is that the user (as root) bumps those
host wide limits.

Extend the script to detect that, so that it can run.

If you now want to run those tests, first issue a

  echo 4096000 > /proc/sys/net/core/rmem_max
  echo 4096000 > /proc/sys/net/core/wmem_max
  echo 4096000 > /proc/sys/net/core/wmem_default

as root. You can get this simply by running:

  $ sudo ./contrib/fw-in-container/fw-in-container -S