fix(translation): fix bogus format string in "po/sl.po"
Some texts are used via str.format(). It's important that even for
translated strings, the names of the replacement fields still match.

Otherwise, with `LANG=sl_SI.UTF-8 firewall-applet` we hit an exception
at

        binding = _("{entry} (Zone: {zone})")
        ...
                action = QAction(
                    escape(binding.format(zone=zone, entry=connection_name)), self
                )

Translations in this sense become part of the program code itself. They
cannot just contain bogus formats. Instead, it's a bug if a translation
mangles the format string. Optimally, _() (gettext.gettext()) would
notice this and ignore the translation, but it doesn't.

https://github.com/firewalld/firewalld/issues/1310

Fixes: ce6d81d7e424 ('chore(translation): update (Slovenian) via Weblate')