improvement(firewalld_conf): track D-Bus metadata for properties in firewalld_conf
Previously, the key meta data for firewalld.conf and D-Bus was separate.
While you might think that this is better, because then firewalld_conf
is independent from the D-Bus API, it is not.

It is worse, because the firewalld.conf entries are exposed on D-Bus.
And that is not some happenstance but something fundamental. Code
doesn't get simpler, by pretending those things would be separate.

The benefit is, when now you review the code to see what keys exist, you
also can see how this is exposed on D-Bus. It is not detached at some
remote place. Likewise, when adding a new key, it becomes immediately
clear, that you also must consider how this will be exposed on D-Bus.

And really, firewalld_conf still has no relevant logic about D-Bus. It
merely has a bit of D-Bus related meta data telling how this property
would look on D-Bus. There is now `import dbus` in "firewalld_conf.py",
but only because "dbus_type=dbus.String" is more useful than an
alternative type annotation like "dbus_type='s'" or "dbus_type=str".