improvement(fw-in-container): support (re)setting size with "--setup-host"
By default, the "--setup-host" bumps the sysctls to 4000k, which should
allow our tests to pass inside fw-in-container.

  $ sudo ./contrib/fw-in-container/fw-in-container -S

Now, also accept an optional size argument to "-S". With this,

  $ sudo ./contrib/fw-in-container/fw-in-container -S $((8000*1024))

explicitly selects the requested size.

The main use is however

  $ sudo ./contrib/fw-in-container/fw-in-container 0

to reset the value sysctl. In this case, "0" is an alias for "212992",
which is the default on my Fedora machine. The point is to be able to
reset the default.