improvement(firewalld_conf): add convenience accessors to firewalld_conf
Internally, the values are all strings and normalize() ensures that.

Now that we know that the values are normalized, we also know that
a boolean key is strictly "yes" or "no".

Still, it's slightly inconvenient. Add "get_as()" helper, that can
convert those strings to booleans. Note that this only accepts
converting the value, if the keytype is also boolean.