improvement(fw): simplify handling of boolean settings from "firewalld.conf"
Don't re-implement the parsing of the values. Granted, now boolean
settings are normalized to "yes" / "no", but still, let's use `get_as()`
which internally also checks that the requested value is really a
boolean type.

Also, don't access the default values like
config.FALLBACK_CLEANUP_ON_EXIT. The key meta data in firewalld_conf
knows what the default value of a key is. Get it from there.

Note that previously there was code like:

        if self._firewalld_conf.get("IPv6_rpfilter"):
            value = self._firewalld_conf.get("IPv6_rpfilter")

as if the key could be unset. But it cannot. Currently, firewalld_conf
will always set all known keys (possibly to their default). Maybe it
would make sense to keep track whether a key is unset in the file or
whether a value was set explicitly. But that does not (currently) exist.
And having code that depends that it could be, is not useful.