improvement(nftables): let _pkttype_match_fragment() lookup the log_denied value itself
We have many functions that take a log_denied argument. On the other hand,
we have many places that directly call self._fw.get_log_denied().

So you might think, it would be better to avoid accessing the global
get_log_denied() at many places, instead only fetch it at the highest
level and pass it down as arguments. The deeper down in the callstack
we are, the less we want to access the global variable.

However, if you fallow that through, you will see that a large amount
of functions will need to get a new log_denied argument. It's a lot
of churn. Instead, it's fine to call self._fw.get_log_denied(), because
that is just a getter (without side effects) of a variable that only changes
during reload().

Hence, let _pkttype_match_fragment() look up the value itself, when the
caller doesn't have a value at hand.

Also, rename the "pkttype" argument to "log_denied". The argument really
holds a value from LogDenied setting, and it needs special if-checks to
map it to the pkttype for the nft API.