test(log_denied_group): add new test "log denied group"
This test is a copy of "src/tests/regression/gh940.at".

At this point, the test is still identical to "gh940.at" (except
changing the test name and keywords). In the next commit, it will be
adjusted. We add first the original LOG based test, so we can see
how the next commit will adjust the test for NFLOG.