feat(IPv6_rpfilter): support loose-forward rpfilter
This adds a new config value for IPv6_rpfilter, loose-forward. In this
mode the rpfilter occurs only for forwarded packets using the "loose"
algorithm from RFC 3704. This is a significant performance improvement
for end-stations that have a single default route because the RPF check
is completely absent on INPUT.

This new value is NOT compatible with the iptables backend. This is
enforced by configuration checks.