fix(firewall-config): fix rich rule creation
It's impossible to create a rich rule because the Rich_Rule constructor
is incorrecly invoked on lines 3404-3709 of src/firewall-config.in.
Moreover, Rich_Rule attributes are set for its instance, which, due to
the frozen state of Rich_Rule, leads to even more errors.

Basically, the fix is quite simple: just create a dictionary of
attributes and pass them to the Rich_Rule constuctor at once.

Fixes: #1463
Fixes: rhbz2361082
Fixes: RHEL-94646
Fixes: 19b04a33e958 ("chore(rich): make Rich_Rule hashable")
Signed-off-by: Oleksandr Zakharov <x.zakharov@proton.me>