fix(nftables): ipset: add entries from GLib loop when idle
Sets with a large amount of entries can take a significant time to
apply. Use the GLib mainloop to add them in chunks when the loop is
idle. This allows dbus calls in between the chunks as the dbus
events/calls have higher priority.

Fixes: #1277
(cherry picked from commit 3874bafc427139e647829d7662577567343aceb6)